% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_graphs.R
\name{assign_within_layer_pos}
\alias{assign_within_layer_pos}
\title{Function for plotting nodes in the network between two layers.}
\usage{
assign_within_layer_pos(N, N_u, input_positions, neuron_spacing)
}
\arguments{
\item{N}{Number of neurons in the first layer.}

\item{N_u}{Number of neurons in the second layer.}

\item{input_positions}{Positions of the neurons in the input layer.}

\item{neuron_spacing}{How much space between the neurons. (do we need this?, seems like igraph overrides or smth)}
}
\value{
Positions of the second layer.
}
\description{
Takes care of the three possible cases. Both layers have even
number of neurons, both layers have odd numbers, or one of each.
}
