% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite_functions.R
\name{coef.LBBNN_Net}
\alias{coef.LBBNN_Net}
\title{Get model coefficients (local explanations) of an LBBNN_Net object}
\usage{
\method{coef}{LBBNN_Net}(
  object,
  dataset,
  inds = NULL,
  output_neuron = 1,
  num_data = 1,
  num_samples = 10,
  ...
)
}
\arguments{
\item{object}{an object of class LBBNN_Net.}

\item{dataset}{Either a train_loader object or a torch_tensor object. In the latter case, the user can supply their own data.}

\item{inds}{a numeric vector of indices indicating which samples in the dataset to be used for local explanations.}

\item{output_neuron}{Which class to explain. In the case where we have more than one output neuron, each one has to be explained separately.}

\item{num_data}{If no indices are chosen, the first num_data of the dataset are automatically used for explanations.}

\item{num_samples}{how many samples to use for model averaging when sampling the weights in the active paths.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The mean coefficients and 95\% CI.
}
\description{
Given an input sample x_1,... x_j (with j the number of variables), the local explanation is found by
considering active paths. If relu activation functions are assumed, each path is a piecewise
linear function, so the contribution for x_j is just the sum of the weights associated with the paths connecting x_j to the output.
The contributions are found by taking the gradient wrt x.
}
\details{
If num_data = 1 (or the user only supplies one sample), the confidence interval is taken around the mean explanation of that sample, using model averaging
of the weights. If num_data > 1, then the confidence interval is obtained wrt all the mean explanations of each individual sample.
}
