% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_activation_functions.R
\name{custom_activation}
\alias{custom_activation}
\title{Generate a custom activation function.}
\usage{
custom_activation()
}
\value{
Returns a `\code{torch::nn_module}, can be used in \code{lbbnn_net}
}
\description{
The first 3 entries are customized
in order to see if we can learn that structure. The rest will be relu.
this function is only for experimental purposes so far.
}
