% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layers.R
\name{density_initialization}
\alias{density_initialization}
\title{Generate initializations for the inclusion parameters.}
\usage{
density_initialization(lower, upper)
}
\arguments{
\item{lower}{numeric scalar.}

\item{upper}{numeric scalar, must be greater than}
}
\value{
A numeric vector of length 2: \code{c(lower,upper)}
}
\description{
Controls the initial density of the network, and thus an important tuning
parameter. The initialization parameters are sampled from 1/(1+exp(-U)),
with U ~ Uniform(lower,upper).
}
\keyword{internal}
