% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOW.R
\name{FLOW}
\alias{FLOW}
\title{Class to generate a FLOW}
\usage{
FLOW(input_dim, transform_type, num_transforms)
}
\arguments{
\item{input_dim}{The dimensionality of each layer. First item is the input vector size.}

\item{transform_type}{The type of transformation. Currently only RNVP is implemented.}

\item{num_transforms}{How many layers of transformations to include in the flow}
}
\description{
Contains an nn_module where the initial vector gets transformed through
all the layers in the module. Also computed the log determinant for the entire
transformation, which is just the sum of the independent layers.
}
\examples{
\donttest{
flow <- FLOW(c(200,100,100),transform_type = 'RNVP',num_transforms = 3)
flow$to(device = 'cpu')
x <- torch::torch_rand(200,device = 'cpu')
output <- flow(x)
z_out <- output$z
print(dim(z_out))
log_det <- output$logdet
print(log_det)
}
}
