% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Gallstone_Dataset}
\alias{Gallstone_Dataset}
\title{Gallstone Dataset}
\format{
This dataset has 319 rows and  38 columns.
}
\source{
\url{https://pmc.ncbi.nlm.nih.gov/articles/PMC11309733/#T2}
}
\usage{
Gallstone_Dataset
}
\description{
Taken from the UCI machine learning repository. The task is
to classify whether the patient had gallstones or not. It contains a mix
of demographic data and bioimpedance data.
}
\keyword{datasets}
