% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataloader.R
\name{get_dataloaders}
\alias{get_dataloaders}
\title{Wrapper around torch dataloader}
\usage{
get_dataloaders(
  dataset,
  train_proportion,
  train_batch_size,
  test_batch_size,
  standardize = TRUE,
  shuffle_train = TRUE,
  shuffle_test = FALSE,
  seed = 1
)
}
\arguments{
\item{dataset}{Some dataset. Must be a data.frame, where the last column is y, the dependent variable.}

\item{train_proportion}{A number between 0 and 1, giving the proportion of data to be used for training.
Usually a large number e.g. 0.8, but could also be smaller than 0.5 even though this is uncommon.}

\item{train_batch_size}{How many samples in each batch in the train dataloader.}

\item{test_batch_size}{How many samples in each batch in the test dataloader.}

\item{standardize}{default is TRUE. Usually an advantage for gradient based optimization.}

\item{shuffle_train}{default is TRUE. Ensures data is randomly shuffled before each iteration.}

\item{shuffle_test}{default is FALSE, as there is no need to shuffle the test data, as the order of the data is irrelevant.}

\item{seed}{The seed used in splitting train/test. Important if one wants to compare to other algorithms that
do not need to use torch dataloaders.}
}
\value{
A list containing a train_loader and a test_loader object.
}
\description{
Takes a dataset and returns both torch train and test dataloaders,
needed for subsequent optimization. Avoids the user having to define these dataloaders
themselves.
}
