% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite_functions.R
\name{get_input_inclusions}
\alias{get_input_inclusions}
\title{Function that checks how many times inputs are included, and from which layer}
\usage{
get_input_inclusions(model)
}
\arguments{
\item{model}{A trained LBBNN model with input_skip.}
}
\value{
A matrix of shape (p, L-1) where p is the number of input variables
and L the total number of layers (including input and output), with each element being 1 if the variable is included
or 0 if not included.
}
\description{
Useful when the number of inputs and/or hidden neurons are very
large, and direct visualization of the network is difficult.
}
