% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBBNN-package.R
\docType{package}
\name{LBBNN-package}
\alias{LBBNN}
\alias{LBBNN-package}
\title{LBBNN: Implementation of Latent Binary Bayesian Neural Networks Using Torch for R}
\description{
Package that implements Latent Binary Bayesian Neural Networks in R. The goal is to implement both the mean-field version, in addition to the version taking into account dependencies between weights, using normalizing flows. In addition, the package will implement input-skip Latent Binary Bayesian Neural Networks.
}
\author{
\strong{Maintainer}: Lars Skaaret-Lund \email{lars.skaaret-lund@nmbu.no}

Authors:
\itemize{
  \item Aliaksandr Hubin \email{aliaksandr.hubin@nmbu.no}
  \item Eirik Høyheim \email{eirik.hoyheim@ffi.no}
}

}
\keyword{internal}
