% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_graphs.R
\name{LBBNN_plot}
\alias{LBBNN_plot}
\title{Function to plot an islab structure after removing weights in non-active paths.}
\usage{
LBBNN_plot(
  model,
  layer_spacing = 1,
  neuron_spacing = 1,
  vertex_size = 10,
  label_size = 0.5,
  edge_width = 0.5
)
}
\arguments{
\item{model}{A trained LBBNN model with input_skip.}

\item{layer_spacing}{Spacing inbetween layers.}

\item{neuron_spacing}{Spacing between neurons within a layer.}

\item{vertex_size}{Size of the neurons.}

\item{label_size}{The size of the text within neurons.}

\item{edge_width}{Width of the edges connecting neurons.}
}
\description{
Uses igraph to plot. Some issues related to the sizes and spacing.
}
\examples{
\donttest{
sizes <- c(2,3,3,2)
problem <- 'multiclass classification'
inclusion_priors <- c(0.1,0.1,0.1)
std_priors <- c(1.0,1.0,1.0)
inclusion_inits <- matrix(rep(c(-10,10),3), nrow = 2, ncol = 3)
device <- 'cpu'
torch::torch_manual_seed(0)
model <- LBBNN_Net(problem_type = problem, sizes = sizes,
                   prior = inclusion_priors, inclusion_inits = inclusion_inits,
                   input_skip = TRUE, std = std_priors, flow = FALSE,
                   num_transforms = 2, dims = c(200,200), device = device)
model$compute_paths_input_skip()
LBBNN_plot(model, 1, 1, 14, 1)
}
}
