% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mgp_dataset}
\alias{mgp_dataset}
\title{Auto MPG daataset}
\format{
This dataset has 398 rows and 24 columns.
}
\source{
\url{https://archive.ics.uci.edu/dataset/9/auto+mpg}
}
\usage{
mgp_dataset
}
\description{
Taken from the UCI machine learning repository.
}
\keyword{datasets}
