% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mice_Dataset}
\alias{Mice_Dataset}
\title{Mice Dataset}
\format{
This dataset has 1080 rows and  78 columns.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/26111164/}
}
\usage{
Mice_Dataset
}
\description{
Only used for internal testing for now.
}
\keyword{datasets}
