% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite_functions.R
\name{plot.LBBNN_Net}
\alias{plot.LBBNN_Net}
\title{Plot LBBNN objects}
\usage{
\method{plot}{LBBNN_Net}(x, type = c("global", "local"), data = NULL, num_samples = 100, ...)
}
\arguments{
\item{x}{An instance of LBBNN_net.}

\item{type}{'global' or 'local'.}

\item{data}{If local is chosen, one sample must be provided to obtain the explanation. Must be a torch_tensor of shape(1,p).}

\item{num_samples}{How many samples to use for model averaging over the weights in case of local explanations.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plot either the global structure or local explanations.
}
