% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train_validate.R
\name{posterior_predict.LBBNN}
\alias{posterior_predict.LBBNN}
\title{Custom posterior_predict function for LBBNNs.}
\usage{
posterior_predict.LBBNN(
  LBBNN,
  mpm,
  newdata,
  draws,
  device = "cpu",
  link = NULL
)
}
\arguments{
\item{LBBNN}{An instance of a trained LBBNN_net.}

\item{mpm}{To use the median probability model or not.}

\item{newdata}{A torch dataloader containing the variables with which to predict.}

\item{draws}{The number of times to sample from the variational posterior.}

\item{device}{The device to perform the operations on. Default is cpu.}

\item{link}{Link function to apply to the output of the network.}
}
\value{
A matrix of size (draws,N,C), where N is the number of data points in the test_loader,
and C the number of classes. (1 for regression).
}
\description{
Draw from the (variational) posterior predictive distribution.
}
