% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite_functions.R
\name{print.LBBNN_Net}
\alias{print.LBBNN_Net}
\title{Print summary of an \code{LBBNN_Net} object}
\usage{
\method{print}{LBBNN_Net}(x, ...)
}
\arguments{
\item{x}{An object of class \code{LBBNN_Net}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the input \code{x}.
}
\description{
Provides a summary of a trained \code{LBBNN_Net} object.
Includes the model type (input-skip or not), whether normalizing flows
are used, module and sub-module structure, number of trainable parameters, and prior
variance and inclusion probabilities for the weights.
}
