% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_explanations.R
\name{quants}
\alias{quants}
\title{Function to obtain empirical 95\% confidence interval, including the mean}
\usage{
quants(x)
}
\arguments{
\item{x}{numeric vector whose sample quantiles is desired.}
}
\value{
The quantiles in addition to the mean.
}
\description{
Using the built in quantile function to return 95\% confidence interval
}
