% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raisin_dataset}
\alias{raisin_dataset}
\title{Raisins Dataset}
\format{
this data frame has 900 rows and the following 8 columns:
\describe{
\item{Area}{Number of pixels within the boundary}
\item{MajorAxisLength}{Length of the main axis}
\item{MinorAxisLength}{Length of the small axis}
\item{Eccentricity}{Measure of the eccentricity of the ellipse}
\item{ConvexArea}{The number of pixels of the
smallest convex shell of the region formed by the raisin grain}
\item{Extent}{Ratio of the region formed by the
raisin grain to the total pixels in the bounding box}
\item{Perimeter}{distance between the boundaries of the
raisin grain and the pixels around it}
\item{Class}{Kecimen or Besni raisin.}
}
}
\source{
\url{https://archive.ics.uci.edu/dataset/850/raisin}
}
\usage{
raisin_dataset
}
\description{
Ilkay Cinar, Murat Kokl and Sakir Tasdemi(2020)
provide a dataset consisting of 2 varieties of Turkish raisins, with 450
samples of each type. The dataset contains 7 morphological features,
extracted from images taken of the Raisins. The goal is to classify to one of
the two types of Raisins.
}
\keyword{datasets}
