% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite_functions.R
\name{residuals.LBBNN_Net}
\alias{residuals.LBBNN_Net}
\title{Residuals from LBBNN fit}
\usage{
\method{residuals}{LBBNN_Net}(object, type = c("response"), ...)
}
\arguments{
\item{object}{An object of class \code{LBBNN_Net}.}

\item{type}{Currently only 'response' is implemented i.e. y_true - y_predicted.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A numeric vector of residuals (\code{y_true - y_predicted})
}
\description{
Residuals from an object of the \code{LBBNN_Net} class.
}
\examples{
\donttest{
# model <- train_LBBNN(...)  # train or load a model
# residuals(model)
}
}
