% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layers.R
\name{std_prior}
\alias{std_prior}
\title{Generate prior standard deviation probabilities}
\usage{
std_prior(x, device = "cpu")
}
\arguments{
\item{x}{A number greater than 0.}

\item{device}{device to train on}
}
\value{
a numeric to be added to either (out_shape,in_shape) in case of linear layers
or (out_channels,in_channels,kernel0,kernel1) in case of convolutional layers
}
\description{
A function to generate prior standard deviations for
each weight and bias in an LBBNN layer. Currently only the same value for
each weight and bias is allowed.
}
\examples{
std_prior(0.3)
}
