% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite_functions.R
\name{summary.LBBNN_Net}
\alias{summary.LBBNN_Net}
\title{Summary of LBBNN fit}
\usage{
\method{summary}{LBBNN_Net}(object, ...)
}
\arguments{
\item{object}{An object of class LBBNN_Net.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The table of the summary values.
}
\description{
Summary method for objects of the LBBNN_Net class. Note that this function only applies to objects trained with input-skip.
}
\details{
For each layer (aside from the output), the summary contains a column showing the number of connections (paths) for each variable from that layer.
In addition, each layer has a column showing the average inclusion probability associated with each variable. The final column shows the average inclusion probability
across all layers.
}
