% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train_validate.R
\name{validate_LBBNN}
\alias{validate_LBBNN}
\title{Function to validate a trained LBBNN model}
\usage{
validate_LBBNN(LBBNN, num_samples, test_dl, device)
}
\arguments{
\item{LBBNN}{An instance of a trained LBBNN_net to be validated}

\item{num_samples}{The number of samples from the variational posterior to be used for model averaging}

\item{test_dl}{An instance of torch dataloader, containing the validation data}

\item{device}{The device to to validate on}
}
\value{
A list containing accuracy if classification, or loss if regression. In both cases
results are returned for both the full model, and the sparse model, only using weights with
a posterior inclusion probability larger than 0.5. The density is also returned.
}
\description{
Computing metrics for each-mini batch in the validation dataset,
without computing gradients. Model averaging (i.e. num samples > 1) is encouraged
for improved performance.
}
