% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Wine_quality_dataset}
\alias{Wine_quality_dataset}
\title{Internal dataset for testing}
\format{
An object of class \code{data.frame} with 6497 rows and 12 columns.
}
\usage{
Wine_quality_dataset
}
\description{
Internal dataset for testing
}
\keyword{internal}
