% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layers.R
\name{alpha_prior}
\alias{alpha_prior}
\title{Generate prior inclusion probabilities for a LBBNN layer}
\usage{
alpha_prior(x, out_shape, in_shape, device)
}
\arguments{
\item{x}{A number between 0 and 1.}

\item{out_shape}{number of output neurons.}

\item{in_shape}{number of input neurons.}
}
\value{
a torch tensor with shape (out_shape,in_shape) with prior probabilities.
}
\description{
A function to generate prior probability values for
each weight in an LBBNN layer. Currently only the same probability for each
weight in a layer is allowed.
}
\examples{
alpha_prior(0.25, 10,5)
}
