% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layers.R
\name{alpha_prior}
\alias{alpha_prior}
\title{Generate prior inclusion probabilities for a LBBNN layer}
\usage{
alpha_prior(x, device = "cpu")
}
\arguments{
\item{x}{A number between 0 and 1.}

\item{device}{The device to be trained on}
}
\value{
a numeric to be added to either (out_shape,in_shape) in case of linear layers
or (out_channels,in_channels,kernel0,kernel1) in case of convolutional layers
}
\description{
A function to generate prior probability values for
each weight in an LBBNN layer. Currently only the same probability for each
weight in a layer is allowed.
}
\examples{
alpha_prior(0.25)
}
