% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layers.R
\name{density_initialization}
\alias{density_initialization}
\title{Function to get initialization values for the inclusion parameters}
\usage{
density_initialization(lower, upper, device = "cpu")
}
\arguments{
\item{lower}{a number}

\item{upper}{a number, greater than lower}

\item{device}{device to train on}
}
\value{
A vector of size two c(lower,upper)
}
\description{
Generate initialization for the inclusion parameters.
This will control the initial density of the network, and thus an important tuning
parameter. The initialization parameters are sampled from 1/(1+exp(-U)),
with U ~ Uniform(lower,upper)
}
\examples{
density_initialization(0,1)
}
