% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_explanations.R
\name{get_local_explanations_gradient}
\alias{get_local_explanations_gradient}
\title{Function to get gradient based local explanations}
\usage{
get_local_explanations_gradient(
  model,
  input_data,
  num_samples = 1,
  magnitude = TRUE,
  include_potential_contribution = FALSE,
  device = "cpu"
)
}
\arguments{
\item{model}{Input-skip model to explain. Not implemented for regular LBBNNs.}

\item{input_data}{The data to be explained (one sample).}

\item{num_samples}{How many sample to use to produce credible intervals.}

\item{magnitude}{If TRUE, only return explanations. If FALSE, multiply by input values.}

\item{include_potential_contribution}{IF TRUE, If covariate=0,
we assume that the contribution is negative (good/bad that it is not included)
if FALSE, just removes zero covariates.}

\item{device}{the device to be trained on. Default is cpu.}
}
\value{
a list containing the losses and accuracies (if classification) and density for each epoch during training.
For comparisons sake we show the density with and without active paths.
}
\description{
Function to get gradient based local explanations
}
