% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBBNN-package.R
\docType{package}
\name{LBBNN-package}
\alias{LBBNN}
\alias{LBBNN-package}
\title{LBBNN: Implementation of Latent Binary Bayesian Neural Networks using torch for R}
\description{
LBBNN package in R
}
\author{
\strong{Maintainer}: Lars Skaaret-Lund \email{lars.skaaret-lund@nmbu.no}

}
\keyword{internal}
