% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNVP.R
\name{MLP}
\alias{MLP}
\title{Multi layer-perceptron}
\usage{
MLP(hidden_sizes, device = "cpu")
}
\value{
Returns a \code{torch::nn_module} representing the MLP.
The module has the following method:
\describe{
\item{\code{forward(x)}}{
Applies each linear layer in \code{hidden_sizes} followed by a LeakyReLU activation (except after
the final layer). Returns a \code{torch::torch_tensor} whose last dimension equals the
last element of \code{hidden_sizes}.
}
}
}
\description{
Generate a multi-layer perceptron, used in RNVP transforms.
}
\keyword{internal}
