% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOW.R, R/RNVP.R
\name{MLP}
\alias{MLP}
\title{Class to generate a multi-layer perceptron, used in RNVP transforms.}
\usage{
MLP(hidden_sizes, device = "cpu")

MLP(hidden_sizes, device = "cpu")
}
\arguments{
\item{hidden_sizes}{A vector of ints. The first is the dimensionality of the vector,
to be transformed by RNVP. The subsequent are hidden dimensions in the MLP.}

\item{device}{The device to be used. Default is CPU.}
}
\description{
As of now, each hidden layer (except the last) is followed
by a non-linear transformation.

As of now, each hidden layer (except the last) is followed
by a non-linear transformation.
}
\examples{
net <- MLP(c(50,100,200,400))
x <- torch::torch_rand(50)
out <- net(x)
print(dim(out))
net <- MLP(c(50,100,200,400))
x <- torch::torch_rand(50)
out <- net(x)
print(dim(out))
}
