% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite_functions.R
\name{summary.LBBNN_Net}
\alias{summary.LBBNN_Net}
\title{Summary of LBBNN fit}
\usage{
\method{summary}{LBBNN_Net}(object, ...)
}
\arguments{
\item{object}{An object of class \code{LBBNN_Net}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A \code{data.frame} containing the above information. The function prints a formatted summary to the console.
The returned \code{data.frame} is invisible.
}
\description{
Summary method for objects of the \code{LBBNN_Net} class.
Only applies to objects trained with \code{input_skip = TRUE}.
}
\details{
The returned table combines two types of information:
\itemize{
\item Number of times each input variable is included in the active paths from each layer
(obtained from \code{get_input_inclusions()}).
\item Average inclusion probabilities for each input variable from each layer, including a final
column showing the average across all layers.
}
}
\examples{
\donttest{
# model <- train_LBBNN(...)  # train or load a model
# summary(model)
}
}
