% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train_validate.R
\name{train_LBBNN}
\alias{train_LBBNN}
\title{Function to train an instance of LBBNN_Net}
\usage{
train_LBBNN(
  epochs,
  LBBNN,
  lr,
  train_dl,
  device = "cpu",
  scheduler = NULL,
  sch_step_size = NULL
)
}
\arguments{
\item{epochs}{Total number of epochs to train for, where one epoch is a pass through the entire training dataset (all mini batches).}

\item{LBBNN}{An instance of the LBBNN_Net class, to be trained.}

\item{lr}{The learning rate to be used in the Adam optimizer.}

\item{train_dl}{An instance of torch dataloader, containing the data to be trained.}

\item{device}{the device to be trained on. Default is cpu.}

\item{scheduler}{A torch learning rate scheduler object. Can be used to decay learning rate for better convergence.}

\item{sch_step_size}{Where to decay if using torch::lr_step. E.g. 1000 means learning rate is decayed every 1000 epochs.}
}
\value{
a list containing the losses and accuracy (if classification) and density for each epoch during training.
For comparisons sake we show the density with and without active paths.
}
\description{
Function that for each epoch iterates through each mini-batch, computing
the loss and using back-propagation to update the network parameters.
}
\examples{
x<-torch::torch_randn(1000,10) #generate some data
b <- torch::torch_rand(10)
y <- torch::torch_matmul(x,b)
med <- torch::torch_median(y)
y[y > med] = 1 #change it into binary classification
y[y <= med] = 0
train_data <- torch::tensor_dataset(x,y)
train_loader <- torch::dataloader(train_data,batch_size = 100,shuffle=TRUE)
problem<-'binary classification'
sizes <- c(10,50,1) #10 input variables, one hidden layer of 50 neurons, 1 output
inclusion_priors <-c(0.3,0.9) #one prior probability per weight matrix
inclusion_inits <- matrix(rep(c(-10,10),2),nrow = 2,ncol = 2)
stds <- c(1.0,1.0)
model <- LBBNN_Net(problem,sizes,inclusion_priors,stds,inclusion_inits,flow = FALSE)
output <- train_LBBNN(epochs = 10,LBBNN = model, lr = 0.01,train_dl = train_loader)
}
